+++ spinta/types/model.py
@@ -18,7 +18,7 @@ from spinta.backends.nobackend.components import NoBackend
 from spinta.commands import authorize
 from spinta.commands import check
 from spinta.commands import load
-from spinta.components import Action, Component, PageBy, Page, PageInfo, UrlParams, pagination_enabled
+from spinta.components import Action, PageBy, Page, PageInfo, UrlParams, pagination_enabled
 from spinta.components import Base
 from spinta.components import Context
 from spinta.components import Mode
@@ -26,7 +26,7 @@ from spinta.components import Model
 from spinta.components import Property
 from spinta.core.access import link_access_param
 from spinta.core.access import load_access_param
-from spinta.core.enums import Level
+from spinta.core.enums import Level, load_level, load_status, load_visibility
 from spinta.datasets.components import ExternalBackend
 from spinta.dimensions.comments.helpers import load_comments
 from spinta.dimensions.enum.components import EnumValue
@@ -35,7 +35,7 @@ from spinta.dimensions.enum.helpers import link_enums
 from spinta.dimensions.enum.helpers import load_enums
 from spinta.dimensions.lang.helpers import load_lang_data
 from spinta.dimensions.param.helpers import load_params
-from spinta.exceptions import KeymapNotSet, InvalidLevel
+from spinta.exceptions import KeymapNotSet
 from spinta.exceptions import PropertyNotFound
 from spinta.exceptions import UndefinedEnum
 from spinta.exceptions import UnknownPropertyType
@@ -51,7 +51,6 @@ from spinta.types.namespace import load_namespace_from_name
 from spinta.ufuncs.loadbuilder.components import LoadBuilder
 from spinta.ufuncs.loadbuilder.helpers import page_contains_unsupported_keys, get_allowed_page_property_types
 from spinta.units.helpers import is_unit
-from spinta.utils.enums import enum_by_value
 from spinta.utils.schema import NA
 
 if TYPE_CHECKING:
@@ -91,6 +90,9 @@ def load(
         model.ns.parents(),
     ))
     load_level(model, model.level)
+    load_status(model, model.status)
+    load_visibility(model, model.visibility)
+
     load_model_properties(context, model, Property, data.get('properties'))
 
     # XXX: Maybe it is worth to leave possibility to override _id access?
@@ -152,9 +154,6 @@ def load(
     builder.load_page()
 
     model.params = load_params(context, manifest, model.params)
-
-    # if "/:" in model.name:
-    #     model.name, model.features = model.name.rsplit("/:", 1)
 
     if not model.name.startswith('_') and not model.basename[0].isupper():
         raise Exception(model.basename, "MODEL NAME NEEDS TO BE UPPER CASED")
@@ -287,6 +286,15 @@ def load(
         prop.given.prepare = prop.prepare_given
     load_level(prop, prop.level)
 
+    # todo properties should inherit those from model OR have a default. They shouldn't be empty
+    #  Now they are empty in cases when model gets a default value, for example in inspect - tests/apie/test_inspect.py
+
+    #  todo the bug is probably in the place where the models and properties are written
+    # prop.given.status = prop.status
+    # prop.status = load_enum_type_item(prop, prop.status, Status)
+    load_status(prop, prop.status)
+    load_visibility(prop, prop.visibility)
+
     if data['type'] is None:
         raise UnknownPropertyType(prop, type=data['type'])
     if data['type'] == 'ref' and not commands.identifiable(prop):
@@ -329,25 +337,6 @@ def load(
     prop.given.explicit = prop.explicitly_given if prop.explicitly_given is not None else True
     prop.given.name = prop.given_name
     return prop
-
-
-def load_level(
-    component: Component,
-    given_level: Union[Level, int, str],
-):
-    if given_level:
-        if isinstance(given_level, Level):
-            level = given_level
-        else:
-            if isinstance(given_level, str) and given_level.isdigit():
-                given_level = int(given_level)
-            if not isinstance(given_level, int):
-                raise InvalidLevel(component, level=given_level)
-            level = enum_by_value(component, 'level', Level, given_level)
-    else:
-        level = None
-    component.level = level
-
 
 def _link_prop_enum(
     prop: Property,